separate bsd specific code
