+++ memusage.h
@@ -1,0 +1,8 @@
+#ifndef MEMUSAGE_H
+#define MEMUSAGE_H
+
+int memusage (pid_t pid);
+void memusage_init (void);
+void memusage_close (void);
+
+#endif
